# smart-home-api-doc.yaml
openapi: 3.0.1
info:
  title: Smart Home Application API
  description: A large scale application for controlling and monitoring smart devices in a city.
  version: v0
servers:
  - url: http://localhost:8080
paths:
  /api/property/deny/{propertyId}:
    put:
      tags:
        - property-controller
      operationId: denyProperty
      description: Endpoint for denying a property.
      parameters:
        - name: propertyId
          in: path
          required: true
          description: ID of the property
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Property denied.
          content:
            'text/plain':
              schema:
                type: object
                description: Property denied!
        '404':
          description: Property does not exist.
          content:
            'text/plain':
              schema:
                type: string
                example: Property does not exist!
  /api/property/approve/{propertyId}:
    put:
      tags:
        - property-controller
      operationId: approveProperty
      description: Endpoint for approving a property.
      parameters:
        - name: propertyId
          in: path
          required: true
          description: ID of the property
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Property approved.
          content:
            'text/plain':
              schema:
                type: object
                example: Property approved!
        '404':
          description: Property does not exist.
          content:
            'text/plain':
              schema:
                type: string
                example: Property does not exist!
...
  /api/property/registerProperty:
    post:
      tags:
        - property-controller
      operationId: register_1
      description: Endpoint for registering new property.
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PropertyDTO'
      responses:
        '200':
          description: Property successfully registered! Waiting for admin to approve.
          content:
            'text/plain':
              schema:
                type: string
                example: Property successfully registered! Waiting for admin to approve.
        '400':
          description: Address already exists.
          content:
            'text/plain':
              schema:
                type: string
                example: Address already exists!
        '404':
          description: Error caused by user or town not existing.
          content:
            'text/plain':
              schema:
                type: string
              examples:
                User not existing:
                  value: User does not exist!
                Town not existing:
                  value: Town does not exist!
...
  /api/property/approvedProperties/{username}:
    get:
      tags:
        - property-controller
      operationId: approvedProperties
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
  /api/property/unapprovedProperties:
    get:
      tags:
        - property-controller
      operationId: unapprovedProperties
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
  /api/property/allProperties:
    get:
      tags:
        - property-controller
      operationId: allProperties
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
  /api/property/approvedProperties:
    get:
      tags:
        - property-controller
      operationId: getAllApprovedProperties
      description: Endpoint for retrieving all approved properties.
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PropertyDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        address:
          type: string
        city:
          type: string
        size:
          type: string
        floors:
          type: string
        status:
          type: string
        propertyType:
          type: string
        picture:
          type: string
          format: binary
        owner:
          type: string
...